UI: moved Alert and Popup into real modules
